requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow
keras
Pillow
opencv-python
joblib
requests
